Add files via upload
All of these are 3D modeled by me off of pictures, all of which were modeled when I was around 16 years old as a Junior in High School. All done in Autodesk Inventor, in Version 2023 brought to 2025. I was only able to model for at most 2 hours a day at a normal school week schedule, yet averaged around an hour or so. I have more that I still have to find and figure out, but this is the current portfolio till I figure out how to get the other soft wares I used.